pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h-agent"
version = "0.1.0"
description = "OpenAI-powered coding agent harness with modular architecture"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "h-agent team"}
]
keywords = ["agent", "openai", "ai", "coding", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
rag = [
    "chromadb>=0.4.0",
]

[project.scripts]
h-agent = "h_agent.cli.commands:main"

[project.urls]
Homepage = "https://github.com/user/h-agent"
Repository = "https://github.com/user/h-agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["h_agent*"]

[tool.setuptools.package-data]
h_agent = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"